\name{TTMap-package}
\alias{TTMap-package}
\alias{TTMap}
\docType{package}
\title{
Two-Tier Mapper: a clustering tool based on 
topological data analysis
}
\description{
TTMap is a clustering method that groups together samples with the 
same deviation in comparison to a control group. It is specially 
useful when the data is small. It is parameter free.
}
\details{
The DESCRIPTION file:
TTMap/DESCRIPTION
Version 1.0
}
\author{
Rachel Jeitziner
Maintainer: Rachel Jeitziner <rachel.jeitziner@epfl.ch>
}
\references{
R. Jeitziner, TTMap, 2017
}
\keyword{TTMap}
\seealso{
rgl, colorRamps
}
\examples{ttmap_part4}